La til tester for pdlservice.hentOgLagreBrukerDataPaBarn
